requirements.txt
setuptools~=76.1.0
streamlit~=1.43.2
pandas~=2.2.3
langchain-core~=0.3.45
langchain~=0.3.20
langchain-google-genai~=2.1.0
langchain-openai~=0.3.9
selenium~=4.29.0
pytest~=8.3.5
pillow~=11.1.0
python-json-logger~=3.3.0
plotly~=6.0.1
pydantic~=2.10.6
groq~=0.19.0
python-dotenv~=1.0.1